pyproject.toml
[project]
name = "agent-migration2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-storage>=3.9.0",
    "langchain>=1.2.9",
    "langchain-openai>=1.1.8",
    "snowflake-connector-python>=4.2.0",
]
